data json

+++ app/Console/Commands/V1/UpdateDataConsumption.php
@@ -51,40 +51,46 @@ class UpdateDataConsumption extends Command
             foreach ($data as $item) {
                 $source_timestamp = Carbon::create($item->source_timestamp);
                 if ($date->diffInDays($source_timestamp) <= 30) {
-                    $decode = bin2hex(base64_decode($item->raw_json));
-                    foreach ($data_frame as $data) {
-                        try {
-                            $split = substr($decode, ($data['start']), ($data['lenght']));
-                            $bin = hex2bin($split);
-                            if ($data['start'] >= 440) {
-                                $json[$data['variable_name']] = (unpack($data['type'], $bin)[1]) / 1000;
-                                $json["data_" . $data['variable_name']] = (unpack($data['type'], $bin)[1]) / 1000;
-                            } else {
-                                if ($data['variable_name'] == "flags") {
-                                    $json[$data['variable_name']] = strval(unpack($data['type'], $bin)[1]);
+                    $pos = strpos(strval($item->raw_json), '{');
+                    if ($pos === false) {
+                        $decode = bin2hex(base64_decode($item->raw_json));
+                        foreach ($data_frame as $data) {
+                            try {
+                                $split = substr($decode, ($data['start']), ($data['lenght']));
+                                $bin = hex2bin($split);
+                                if ($data['start'] >= 440) {
+                                    $json[$data['variable_name']] = (unpack($data['type'], $bin)[1]) / 1000;
+                                    $json["data_" . $data['variable_name']] = (unpack($data['type'], $bin)[1]) / 1000;
                                 } else {
-                                    $json[$data['variable_name']] = unpack($data['type'], $bin)[1];
+                                    if ($data['variable_name'] == "flags") {
+                                        $json[$data['variable_name']] = strval(unpack($data['type'], $bin)[1]);
+                                    } else {
+                                        $json[$data['variable_name']] = unpack($data['type'], $bin)[1];
+                                    }
                                 }
-                            }
 
-                            if (is_nan($json[$data['variable_name']])) {
-                                $json[$data['variable_name']] = null;
-                            }
+                                if (is_nan($json[$data['variable_name']])) {
+                                    $json[$data['variable_name']] = null;
+                                }
 
-                            if ($data['variable_name'] == "ph3_varLh_acumm") {
-                                break;
+                                if ($data['variable_name'] == "ph3_varLh_acumm") {
+                                    break;
+                                }
+                            } catch (Exception $e) {
+                                echo 'Excepción capturada: ', $e->getMessage(), "\n";
                             }
-                        } catch (Exception $e) {
-                            echo 'Excepción capturada: ', $e->getMessage(), "\n";
                         }
+                        $item->raw_json = $json;
+                        if ($json['import_wh'] <= 0) {
+                            $item->updateQuietly();
+                            $item->delete();
+                            return;
+                        }
+                    } else {
+                        $source_timestamp->addMinute();
+                        $item->source_timestamp = $source_timestamp->format("Y-m-d H:i:s");
                     }
-                    $item->raw_json = $json;
-                    if ($json['import_wh'] <= 0) {
-                        $item->updateQuietly();
-                        $item->delete();
-                        return;
-                    }
-                    $item->update();
+                    $item->save();
                 } else{
                     $item->delete();
                 }
